Add a way to register custom resolvers for apollo from a render app
Right now it's impossible for render apps to add any configuration to the apollo client. Some configurations can be important for the apps though, such as registering custom resolvers to reuse cached data between queries. This is app specific, so we can't figure out what the user would want, instead we need to let them configure this on their own.

For this I've just exposed some functions to allow this feature, and made sure that the apollo client is only created when used for the first time (to make sure all code that may be registering custom resolvers will already have run).