Export function to listen to browser history changes
The browser history object built by the `Route` component is hard to access, even though it's stored globally, because it's created after nested components have already finished rendering.

This function is a way for app components to listen to history changes, while still sharing the same browser history object.

In the future it would be a lot better if we could have all location info be accessible in a graphql way (our long wished local graphql), so that components that need this info can automatically be updated in the same way as with any other data.

Until we have that though, we need something like this (I think, let me know if anyone has a better idea).