Change apollo's `dataIdFromObject` config to only handle `id` fields specially
We want this logic to be generic for all masterdata schemas, so this is removing the specific id logic for `slug` and other fields, and keeping only the pattern of having `${__typename}:${id}`. To have data automatically cached in the client, users will need to provide an `id` field for their types.