Change missing container error log to warn
Turns out this is a possible scenario. If we add a new LegacyExtensionContainer child (a new portal to be rendered, like checkout/payments) and the markup is not included in page's HTML, because the Portal Template did not include the viewpart for this extension yet, this portal container will be missing.